keycloak/requirements.txt
python-keycloak>=3.9.0
requests>=2.31.0
python-dotenv>=1.0.0
